requirements.txt
pytest
PyYaml>=5.1
requests>=2.0
rich
